main/livekit-media/requirements.txt
# LiveKit SDK
livekit-rtc>=0.12.0
livekit-api>=0.6.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Qdrant (Simple client, no ML dependencies)
qdrant-client>=1.7.0

# HTTP Client
aiohttp>=3.9.0

# Audio Processing
pydub>=0.25.1

# Environment
python-dotenv>=1.0.0
